requirements.txt
pypdf
langchain
faiss-cpu
sentence_transformers
openai
python-dotenv
pytube
nemoguardrails